requirements.txt
flask==2.0.1
werkzeug==2.0.3
requests==2.32.3
python-dotenv==1.0.0 